Fixed the option to give Cornelius 2000 gold constantly and simplified the dialogue with him by removing dialogue options.

+++ CONTENT/Story/Dialoge/DIA_VLK_401_Cornelius.d
@@ -53,31 +53,7 @@ FUNC VOID DIA_Cornelius_SeeMurder_Info()
 	AI_Output (other,self ,"DIA_Cornelius_SeeMurder_15_00"); //You saw the murder of Lothar, didn't you?
 	AI_Output (self ,other,"DIA_Cornelius_SeeMurder_13_01"); //(nervously) I need not answer any questions about a current investigation.
 	AI_Output (self ,other,"DIA_Cornelius_SeeMurder_13_02"); //Lord Hagen already has everything I know on record.
-};
 
-//*********************************************************************
-//	Was hast du gesehen? 
-//********************************************************************
-INSTANCE DIA_Cornelius_WhatYouSee   (C_INFO)
-{
-	npc         = VLK_401_Cornelius;
-	nr          = 5;
-	condition   = DIA_Cornelius_WhatYouSee_Condition;
-	information = DIA_Cornelius_WhatYouSee_Info;
-	permanent   = FALSE;
-	description	= "What was it you saw?";
-};
-
-FUNC INT DIA_Cornelius_WhatYouSee_Condition()
-{
-	if Npc_KnowsInfo (other,DIA_Cornelius_SeeMurder)
-	{
-		return TRUE;
-	};	
-};
-
-FUNC VOID DIA_Cornelius_WhatYouSee_Info()
-{
 	AI_Output (other,self ,"DIA_Cornelius_WhatYouSee_15_00"); //What was it you saw?
 	AI_Output (self ,other,"DIA_Cornelius_WhatYouSee_13_01"); //(hectically) Really, I haven't got any time now.
 	AI_Output (self ,other,"DIA_Cornelius_WhatYouSee_13_02"); //(hectically) You need to go now, the office is closing.
@@ -102,7 +78,7 @@ INSTANCE DIA_Cornelius_Enough   (C_INFO)
 
 FUNC INT DIA_Cornelius_Enough_Condition()
 {
-	if Npc_KnowsInfo (other,DIA_Cornelius_WhatYouSee)
+	if Npc_KnowsInfo (other,DIA_Cornelius_SeeMurder)
 	{
 		return TRUE;
 	};	
@@ -148,22 +124,9 @@ FUNC VOID DIA_Cornelius_DontBelieveYou_Info()
 	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_13_01"); //(with fake confidence) So what, what are you going to do about it?
 
 	Info_ClearChoices (DIA_Cornelius_DontBelieveYou);
-	//hier fehlt noch Taschendiebstahl
-	Info_AddChoice (DIA_Cornelius_DontBelieveYou,"What are you asking?",DIA_Cornelius_DontBelieveYou_WhatYouWant);
+
+	Info_AddChoice (DIA_Cornelius_DontBelieveYou,"What is it you want?",DIA_Cornelius_DontBelieveYou_WhatYouWant);
 	Info_AddChoice (DIA_Cornelius_DontBelieveYou,"You hold your life dear, don't you?",DIA_Cornelius_DontBelieveYou_WantSurvive);
-	
-	if (hero.guild == GIL_KDF)
-	{
-		Info_AddChoice (DIA_Cornelius_DontBelieveYou,"They could make you talk in the monastery.",DIA_Cornelius_DontBelieveYou_Monastery);
-	};
-	if (hero.guild == GIL_SLD)
-	{
-		Info_AddChoice (DIA_Cornelius_DontBelieveYou,"I could tell the mercenaries where you live.",DIA_Cornelius_DontBelieveYou_KnowYourHome);
-	};
-	if (hero.guild == GIL_MIL)
-	{
-		Info_AddChoice (DIA_Cornelius_DontBelieveYou,"Perjury will land you in the slammer - for a long time!",DIA_Cornelius_DontBelieveYou_Perjury);
-	};
 };
 
 FUNC VOID DIA_Cornelius_DontBelieveYou_WhatYouWant ()
@@ -186,42 +149,36 @@ FUNC VOID DIA_Cornelius_DontBelieveYou_WantSurvive ()
 	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_WantSurvive_13_02"); //I have powerful friends. So don't you dare lay a finger on me.
 	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_WantSurvive_13_03"); //Now get out of here! Or I'll call the guards!
 	
-	
-	
-	AI_StopProcessInfos (self);
-};
-
-FUNC VOID DIA_Cornelius_DontBelieveYou_Monastery()
-{
-	AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_Monastery_15_00"); //They could make you talk in the monastery.
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Monastery_13_01"); //(white as chalk) What do you mean by that?
-	AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_Monastery_15_02"); //Well, we have ways to bring out the truth. Painful ways.
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Monastery_13_03"); //No, please don't. I'll tell you everything you want.
-	
-	Cornelius_TellTruth = TRUE;
-	Info_ClearChoices (DIA_Cornelius_DontBelieveYou);
-};
-
-FUNC VOID DIA_Cornelius_DontBelieveYou_KnowYourHome()
-{
-	AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_KnowYourHome_15_00"); //I could tell the mercenaries where you live.
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_KnowYourHome_13_01"); //(white as chalk) What is that supposed to mean?
-	AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_KnowYourHome_15_02"); //I bet they're dying to make your acquaintance. They are not very happy at all.
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_KnowYourHome_13_03"); //You can't do that, they would kill me.
-	AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_KnowYourHome_15_04"); //That is very well possible.
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_KnowYourHome_13_05"); //I'll say whatever you want, but you mustn't do that.
-	
-	Cornelius_TellTruth = TRUE;
-	Info_ClearChoices (DIA_Cornelius_DontBelieveYou);
-};
-
-FUNC VOID DIA_Cornelius_DontBelieveYou_Perjury()
-{
-	AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_Perjury_15_00"); //Perjury will land you in the slammer - for a long time!
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Perjury_13_01"); //You're trying to threaten me? A little guardsman is threatening me, the secretary to the governor?
-	AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Perjury_13_02"); //If you don't get lost right now, I'll see to it that you are demoted.
-	
-	Cornelius_ThreatenByMilSC = TRUE;
+	if (hero.guild == GIL_KDF)
+	{
+		AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_Monastery_15_00"); //They could make you talk in the monastery.
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Monastery_13_01"); //(white as chalk) What do you mean by that?
+		AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_Monastery_15_02"); //Well, we have ways to bring out the truth. Painful ways.
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Monastery_13_03"); //No, please don't. I'll tell you everything you want.
+		
+		Cornelius_TellTruth = TRUE;
+		Info_ClearChoices (DIA_Cornelius_DontBelieveYou);
+	};
+	if (hero.guild == GIL_SLD)
+	{
+		AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_KnowYourHome_15_00"); //I could tell the mercenaries where you live.
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_KnowYourHome_13_01"); //(white as chalk) What is that supposed to mean?
+		AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_KnowYourHome_15_02"); //I bet they're dying to make your acquaintance. They are not very happy at all.
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_KnowYourHome_13_03"); //You can't do that, they would kill me.
+		AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_KnowYourHome_15_04"); //That is very well possible.
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_KnowYourHome_13_05"); //I'll say whatever you want, but you mustn't do that.
+		
+		Cornelius_TellTruth = TRUE;
+		Info_ClearChoices (DIA_Cornelius_DontBelieveYou);
+	};
+	if (hero.guild == GIL_MIL)
+	{
+		AI_Output (other,self ,"DIA_Cornelius_DontBelieveYou_Perjury_15_00"); //Perjury will land you in the slammer - for a long time!
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Perjury_13_01"); //You're trying to threaten me? A little guardsman is threatening me, the secretary to the governor?
+		AI_Output (self ,other,"DIA_Cornelius_DontBelieveYou_Perjury_13_02"); //If you don't get lost right now, I'll see to it that you are demoted.
+		
+		Cornelius_ThreatenByMilSC = TRUE;
+	};
 	
 	AI_StopProcessInfos (self);
 };
@@ -241,7 +198,8 @@ INSTANCE DIA_Cornelius_PayCornelius   (C_INFO)
 
 FUNC INT DIA_Cornelius_PayCornelius_Condition()
 {
-	if Cornelius_PayForProof == TRUE
+	if (Cornelius_PayForProof == TRUE)
+	&& (Cornelius_TellTruth != TRUE)
 	&& (Npc_HasItems (other,ItMi_Gold) >= 2000)
 	{
 		return TRUE;
@@ -252,7 +210,7 @@ FUNC VOID DIA_Cornelius_PayCornelius_Info()
 {
 	AI_Output (other,self ,"DIA_Cornelius_PayCornelius_15_00"); //Here's the gold.
 	
-	B_GiveInvItems (other,self ,ItMi_Gold,2000);
+	B_GiveInvItems (other, self, ItMi_Gold, 2000);
 	
 	AI_Output (self ,other,"DIA_Cornelius_PayCornelius_13_01"); //It's better if I don't ask you where you got that.
 	AI_Output (self ,other,"DIA_Cornelius_PayCornelius_13_02"); //To be honest, I don't really care.
